feat: add `count` option
This commit add new `count` option to control the number of screen
column to display. The purpose of this option is to separate the
screen column configuration and the screen column display
configuration. This allow users to have dynamic numbers of screen
column displayed while keeping their screen column configuration
unchanged.

The defaut value for `count` is `-1` which mean to use the maximum
number of screen column as specified in the `virtcolumn` option.
The value of `count` is clamped the maximum number of screen
column defined in the `virtcolumn` option. Setting the value `0` to
`count` effectively disables rendering all the screen columns.

The displayed screen columns are taken from `virtcolumn` option
indexed from left-to-right in ascending order.

Signed-off-by: Jakka Prihatna <jprihatna@gmail.com>